Add histogram to image layer
