+++ src/neuroglancer/widget/histogram.css
@@ -1,0 +1,4 @@
+.neuroglancer-invlerp-histogram-panel {
+  height: 100px;
+  border: 1px solid #666;
+}
